requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=1.4.0=pypi_0
aiohttp=3.8.4=pypi_0
aiosignal=1.3.1=pypi_0
antlr4-python3-runtime=4.9.3=pypi_0
anyio=3.6.2=pypi_0
argcomplete=3.0.8=pypi_0
argon2-cffi=21.3.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
arrow=1.2.3=pypi_0
asttokens=2.2.1=pypi_0
async-timeout=4.0.2=pypi_0
attrs=23.1.0=pypi_0
backcall=0.2.0=pypi_0
beautifulsoup4=4.12.2=pypi_0
bitarray=2.7.3=pypi_0
blas=1.0=mkl
bleach=6.0.0=pypi_0
boto=2.49.0=pypi_0
brotlipy=0.7.0=py310h7f8727e_1002
bzip2=1.0.8=h7b6447c_0
ca-certificates=2023.01.10=h06a4308_0
cachetools=5.3.0=pypi_0
certifi=2022.12.7=py310h06a4308_0
cffi=1.15.1=py310h5eee18b_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
cloudpickle=2.2.1=pypi_0
colorama=0.4.6=pypi_0
comm=0.1.3=pypi_0
contourpy=1.0.7=pypi_0
cpuonly=2.0=0
crcmod=1.7=pypi_0
cryptography=39.0.1=py310h9ce1e76_0
cvxpy=1.3.1=pypi_0
cycler=0.11.0=pypi_0
cython=0.29.35=pypi_0
datasets=2.12.0=pypi_0
debugpy=1.6.7=pypi_0
decorator=5.1.1=pypi_0
defusedxml=0.7.1=pypi_0
dill=0.3.6=pypi_0
docopt=0.6.2=pypi_0
ecos=2.0.12=pypi_0
executing=1.2.0=pypi_0
fasteners=0.18=pypi_0
fastjsonschema=2.16.3=pypi_0
ffmpeg=4.3=hf484d3e_0
filelock=3.9.0=py310h06a4308_0
fonttools=4.39.4=pypi_0
fqdn=1.5.1=pypi_0
freetype=2.12.1=h4a9f257_0
frozenlist=1.3.3=pypi_0
fsspec=2023.5.0=pypi_0
gcs-oauth2-boto-plugin=3.0=pypi_0
giflib=5.2.1=h5eee18b_3
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=py310heeb90bb_0
gnutls=3.6.15=he1e5248_0
google-apitools=0.5.32=pypi_0
google-auth=2.17.3=pypi_0
google-auth-oauthlib=1.0.0=pypi_0
google-reauth=0.1.1=pypi_0
grpcio=1.54.0=pypi_0
gsutil=5.23=pypi_0
gym=0.26.2=pypi_0
gym-notices=0.0.8=pypi_0
httplib2=0.20.4=pypi_0
huggingface-hub=0.14.1=pypi_0
hydra-core=1.3.2=pypi_0
idna=3.4=py310h06a4308_0
intel-openmp=2023.1.0=hdb19cb5_46305
ipykernel=6.23.0=pypi_0
ipython=8.13.2=pypi_0
ipython-genutils=0.2.0=pypi_0
isoduration=20.11.0=pypi_0
jedi=0.18.2=pypi_0
jinja2=3.1.2=py310h06a4308_0
joblib=1.2.0=pypi_0
jpeg=9e=h5eee18b_1
jsonpointer=2.3=pypi_0
jsonschema=4.17.3=pypi_0
jupyter-client=8.2.0=pypi_0
jupyter-core=5.3.0=pypi_0
jupyter-events=0.6.3=pypi_0
jupyter-server=2.5.0=pypi_0
jupyter-server-terminals=0.4.4=pypi_0
jupyterlab-pygments=0.2.2=pypi_0
kiwisolver=1.4.4=pypi_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
learn2learn=0.2.0=pypi_0
lerc=3.0=h295c915_0
libdeflate=1.17=h5eee18b_0
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.2=h7f8727e_0
libpng=1.6.39=h5eee18b_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.0=h6a678d5_2
libunistring=0.9.10=h27cfd23_0
libuuid=1.41.5=h5eee18b_0
libwebp=1.2.4=h11a3e52_1
libwebp-base=1.2.4=h5eee18b_1
lightning-utilities=0.8.0=pypi_0
lxml=4.9.2=pypi_0
lz4-c=1.9.4=h6a678d5_0
markdown=3.4.3=pypi_0
markupsafe=2.1.1=py310h7f8727e_0
matplotlib=3.7.1=pypi_0
matplotlib-inline=0.1.6=pypi_0
mistune=2.0.5=pypi_0
mkl=2023.1.0=h6d00ec8_46342
mkl-service=2.4.0=py310h5eee18b_1
mkl_fft=1.3.6=py310h1128e8f_1
mkl_random=1.2.2=py310h1128e8f_1
monotonic=1.6=pypi_0
mosek=10.0.44=pypi_0
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpmath=1.2.1=pypi_0
multidict=6.0.4=pypi_0
multiprocess=0.70.14=pypi_0
nbclassic=1.0.0=pypi_0
nbclient=0.7.4=pypi_0
nbconvert=7.4.0=pypi_0
nbformat=5.8.0=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.5.6=pypi_0
nettle=3.7.3=hbbd107a_1
networkx=2.8.4=py310h06a4308_1
notebook=6.5.4=pypi_0
notebook-shim=0.2.3=pypi_0
numpy=1.24.3=py310h5f9d8c6_1
numpy-base=1.24.3=py310hb5e798b_1
oauth2client=4.1.3=pypi_0
oauthlib=3.2.2=pypi_0
omegaconf=2.3.0=pypi_0
openh264=2.1.1=h4ff587b_0
openssl=1.1.1t=h7f8727e_0
osqp=0.6.2.post9=pypi_0
packaging=23.1=pypi_0
pandas=2.0.1=pypi_0
pandocfilters=1.5.0=pypi_0
parso=0.8.3=pypi_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=9.4.0=py310h6a678d5_0
pip=23.0.1=py310h06a4308_0
pipreqs=0.4.13=pypi_0
platformdirs=3.5.0=pypi_0
poisson-disc=0.2.1=pypi_0
portalocker=2.7.0=pypi_0
prometheus-client=0.16.0=pypi_0
prompt-toolkit=3.0.38=pypi_0
protobuf=4.23.0=pypi_0
psutil=5.9.5=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.2=pypi_0
pyarrow=12.0.0=pypi_0
pyasn1=0.5.0=pypi_0
pyasn1-modules=0.3.0=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.15.1=pypi_0
pyopenssl=23.0.0=py310h06a4308_0
pyparsing=3.0.9=pypi_0
pyrsistent=0.19.3=pypi_0
pysocks=1.7.1=py310h06a4308_0
python=3.10.11=h7a1cb2a_2
python-dateutil=2.8.2=pypi_0
python-json-logger=2.0.7=pypi_0
pytorch=2.0.1=py3.10_cpu_0
pytorch-lightning=2.0.2=pypi_0
pytorch-mutex=1.0=cpu
pytz=2023.3=pypi_0
pyu2f=0.1.5=pypi_0
pyyaml=6.0=pypi_0
pyzmq=25.0.2=pypi_0
qdldl=0.1.7=pypi_0
qpth=0.0.15=pypi_0
readline=8.2=h5eee18b_0
regex=2023.5.5=pypi_0
requests=2.29.0=py310h06a4308_0
requests-oauthlib=1.3.1=pypi_0
responses=0.18.0=pypi_0
retry-decorator=1.1.1=pypi_0
rfc3339-validator=0.1.4=pypi_0
rfc3986-validator=0.1.1=pypi_0
rpy2=3.5.11=pypi_0
rsa=4.7.2=pypi_0
sacrebleu=2.3.1=pypi_0
scikit-learn=1.2.2=pypi_0
scipy=1.10.1=pypi_0
scs=3.2.3=pypi_0
send2trash=1.8.2=pypi_0
setuptools=66.0.0=py310h06a4308_0
six=1.16.0=pypi_0
sniffio=1.3.0=pypi_0
soupsieve=2.4.1=pypi_0
sqlite=3.41.2=h5eee18b_0
stack-data=0.6.2=pypi_0
sympy=1.11.1=py310h06a4308_0
tabulate=0.9.0=pypi_0
tbb=2021.8.0=hdb19cb5_0
tensorboard=2.13.0=pypi_0
tensorboard-data-server=0.7.0=pypi_0
terminado=0.17.1=pypi_0
threadpoolctl=3.1.0=pypi_0
tinycss2=1.2.1=pypi_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.13.3=pypi_0
torchaudio=2.0.2=py310_cpu
torchmetrics=0.11.4=pypi_0
torchvision=0.15.2=py310_cpu
tornado=6.3.1=pypi_0
tqdm=4.65.0=pypi_0
traitlets=5.9.0=pypi_0
transformers=4.29.1=pypi_0
typing_extensions=4.5.0=py310h06a4308_0
tzdata=2023.3=pypi_0
tzlocal=5.0=pypi_0
uri-template=1.2.0=pypi_0
urllib3=1.26.15=py310h06a4308_0
wcwidth=0.2.6=pypi_0
webcolors=1.13=pypi_0
webencodings=0.5.1=pypi_0
websocket-client=1.5.1=pypi_0
werkzeug=2.3.4=pypi_0
wheel=0.38.4=py310h06a4308_0
xxhash=3.2.0=pypi_0
xz=5.4.2=h5eee18b_0
yarg=0.1.9=pypi_0
yarl=1.9.2=pypi_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.5=hc292b87_0

numpy~=1.24.3
matplotlib~=3.7.1
cvxpy~=1.3.1
pandas~=2.0.1
mosek~=10.0.44
rpy2~=3.5.11
scikit-learn~=1.2.2
transformers~=4.29.1
datasets~=2.12.0
learn2learn~=0.2.0